Simplify responses polling endpoints
