+++ packages/supercompat/src/adapters/storage/responsesStorageAdapter/threads/run/get.ts
@@ -1,5 +1,6 @@
 import type OpenAI from 'openai'
 import { runRegexp } from '@/lib/runs/runRegexp'
+import { serializeResponseAsRun } from '@/lib/responses/serializeResponseAsRun'
 
 type GetResponse = Response & {
   json: () => Promise<ReturnType<OpenAI.Beta.Threads.Runs['retrieve']>>
@@ -7,24 +8,27 @@ type GetResponse = Response & {
 
 export const get = ({
   openai,
+  openaiAssistant,
 }: {
   openai: OpenAI
+  openaiAssistant: OpenAI.Beta.Assistants.Assistant
 }) => async (urlString: string): Promise<GetResponse> => {
   const url = new URL(urlString)
 
-  const [, threadId, runId] = url.pathname.match(new RegExp(runRegexp))!
+  const [, _threadId, runId] = url.pathname.match(new RegExp(runRegexp))!
 
-  throw new Error('Not implemented')
+  const response = await openai.responses.retrieve(runId)
 
+  const data = serializeResponseAsRun({
+    response,
+    assistantId: openaiAssistant.id,
+  })
 
-  // return
-  // return new Response(JSON.stringify(
-  //   serializeRun({ run })
-  // ), {
-  //   status: 200,
-  //   headers: {
-  //     'Content-Type': 'application/json',
-  //     'openai-poll-after-ms': '5000',
-  //   },
-  // })
+  return new Response(JSON.stringify(data), {
+    status: 200,
+    headers: {
+      'Content-Type': 'application/json',
+      'openai-poll-after-ms': '5000',
+    },
+  })
 }
