+++ packages/supercompat/src/adapters/storage/responsesStorageAdapter/threads/run/index.ts
@@ -6,11 +6,13 @@ import type { RequestHandler } from '@/types'
 
 export const run = ({
   openai,
+  openaiAssistant,
   runAdapter,
 }: {
   openai: OpenAI
+  openaiAssistant: OpenAI.Beta.Assistants.Assistant
   runAdapter: RunAdapter
 }): { get: RequestHandler } => ({
-  get: get({ openai }),
+  get: get({ openai, openaiAssistant }),
   // post: post({ prisma, runAdapter }),
 })
